make dupes disable their lights only if light is sufficient for lit workspace
It seems weird that a dupe would turn off the personal light and reduce
the existing light below what is required for the Lit Workspace bonus.
This is actually exactly what happens with the Ceiling Light placed
at the ceiling of a 4-tile-high room, since it gives 900 lux at the bottom
tile where it matters, and a dupe turning off the personal 200 lux just
makes the difference for the 1000 lux required for the Lit Workspace bonus.